separated createState and createLightState to prevent tree-shaking issue
