+++ packages/core/src/state/public/index.ts
@@ -1,0 +1,9 @@
+import { CreateAgileSubInstanceInterface } from '../../shared';
+import { StateConfigInterface } from '../state';
+
+export * from './createState';
+export * from './createLightState';
+
+export interface CreateStateConfigInterfaceWithAgile
+  extends CreateAgileSubInstanceInterface,
+    StateConfigInterface {}
