+++ packages/event/src/index.ts
@@ -1,4 +1,6 @@
 import { Event } from './event';
 
 export * from './event';
+export * from './react';
+
 export default Event;
